Move tests for `ReactionVariationsUtils` to dedicated file
